pyproject.toml
[project]
name = "eeg-seizure-analyzer"
version = "0.1.0"
description = "Detect and quantify seizures and interictal activity in mouse EEG recordings"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pyedflib>=0.1.34",
    "streamlit>=1.30",
    "plotly>=5.18",
    "pandas>=2.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
windows = ["adi-reader>=0.0.13", "cffi"]
ml = ["scikit-learn>=1.3", "mne>=1.6"]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
eeg-convert = "eeg_seizure_analyzer.io.adicht_to_edf:main"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
